Update detection_abstract.py
Updated validator for "tests" so that it always runs, even if no tests are provided.  without adding always=True, it fails to catch missing tests.
Also, ensure that we run against the .value of the enumeration, 
not the enum object itself. This is required since The Pydantic config use_enum_values = True