Add dnslookup to list of ignored lookups
The 'dnslookup' command is a standard external lookup provided by default in Splunk Enterprise. This commit adds it to the LOOKUPS_TO_IGNORE set in 'contentctl/objects/lookup.py'. This prevents validation errors during the build process when detections reference this lookup, as it does not need to be packaged with the app.